Refactor wallet dimensions and Gradient extension method

+++ binkapp/Wallet/Join Cards/WalletPrompt.swift
@@ -72,7 +72,7 @@ enum WalletPromptType {
         }
     }
 
-    var numberOfRows: Int {
+    var numberOfRows: CGFloat {
         switch self {
         case .link(let plans):
             return plans.count > 2 ? 2 : 1
@@ -124,7 +124,7 @@ class WalletPrompt: WalletPromptProtocol {
         return type.membershipPlans
     }
     
-    var numberOfRows: Int {
+    var numberOfRows: CGFloat {
         return type.numberOfRows
     }
